refactor: extract SparkLine into its own file and use button for dropdown overlay
Move SparkLine component from TopicsTable.tsx to a dedicated SparkLine.tsx
file for better separation of concerns. Replace the dropdown overlay div
with a button element for improved semantics.
